Move from South to native migrations, get tests working.
* Deleted all unused test files.
* Added `TEST_RUNNER` setting.

* Removed all old South migrations (both .py and .pyc files)
* Remove South from `INSTALLED_APPS` and requirements.
* Ran:

    $ ./manage.py makemigrations
    $ ./manage.py migrate
    $ ./manage.py migrate --fake encyclopedia
    $ ./manage.py migrate

At the third step, Answered 'yes' to removing `auth | user`.
