--- pepysdiary/encyclopedia/tests/__init__.py
@@ -1,1 +1,0 @@
-from topicmanager import *
